refactor(client): split config.rs into directory module
Extract RedirectConfig, TimeoutConfig, and RetryPolicy into
config/types.rs submodule. The main Config struct, parsing logic,
and tests remain in config/mod.rs. Re-exports preserve public API.

Reduces mod.rs from 1066 to 809 lines.
